Fix some compiler warnings

+++ NEngine/include/NEngine/Renderer/Texture.h
@@ -25,8 +25,8 @@ class Texture : public Bindable
 {
 public:
     Texture(Helpers::DeviceResources &deviceResources,
-            unsigned int width,
-            unsigned int height,
+            float width,
+            float height,
             TextureBindTarget bindTarget,
             const std::string &name);
